# server/common/api.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: Way Farer
  description: Way Farer is a public bus transportation booking service.(Project submitted for Andela Developer Challenge)
basePath: /api/v1
host: 'localhost:3000'
tags:
  #######################################################################
  #  - name: Examples
  #    description: Simple example endpoints
  #  - name: Specification
  #    description: The swagger API specification
  #######################################################################

  - name: Authentication
    description: User Authentication endpoints
  - name: Trips
    description: Trip endpoints
  - name: Bookings
    description: Trip booking endpoints

consumes:
  - application/json
produces:
  - application/json

definitions:
  #######################################################################
  #  ExampleBody:
  #    type: object
  #    title: example
  #    required:
  #      - name
  #    properties:
  #      name:
  #        type: string
  #        example: no_stress
  #######################################################################

  UserRegisterBody:
    type: object
    title: register
    required:
      - email
      - password
      - first_name
      - last_name
    properties:
      email:
        type: string
        format: email
        example: user1@site.com
      password:
        type: string
        format: password
        example: password
      first_name:
        type: string
        example: fuser1
      last_name:
        type: string
        example: luser1

  UserLoginBody:
    type: object
    title: login
    properties:
      email:
        type: string
        format: email
        example: user1@site.com
      password:
        type: string
        format: password
        example: password

  TripBody:
    type: object
    title: trip
    required:
      - seating_capacity
      - bus_license_number
      - origin
      - destination
      - trip_date
      - fare
    properties:
      seating_capacity:
        type: integer
        format: int32
        example: 40
      bus_license_number:
        type: string
        example: 111
      origin:
        type: string
        example: A
      destination:
        type: string
        example: B
      trip_date:
        type: string
        format: date
        example: 2019-07-18T00:00:00.000Z
      fare:
        type: number
        example: 1

  BookingBody:
    type: object
    title: booking
    required:
      - trip_id
      - user_id
    properties:
      trip_id:
        type: integer
        format: int32
        example: 1
      user_id:
        type: integer
        format: int32
        example: 1

  ErrorAPIResponse:
    type: object
    properties:
      status:
        type: string
      error:
        type: string

  AuthAPIResponse:
    type: object
    properties:
      status:
        type: string
      data:
        type: object
        properties:
          token:
            type: string
          first_name:
            type: string
          last_name:
            type: string
          email:
            type: string
            format: email

  BookingAPIResponse:
    type: object
    properties:
      status:
        type: string
      data:
        type: object
        properties:
          booking_id:
            type: integer
            format: int32
          bus_license_number:
            type: string
          trip_date:
            type: string
            format: date
          first_name:
            type: string
          last_name:
            type: string
          user_email:
            type: string
            format: email

  BookingDeleteAPIResponse:
    type: object
    properties:
      status:
        type: string
      error:
        type: string

  TripAPIResponse:
    type: object
    properties:
      status:
        type: string
      data:
        type: object
        properties:
          booking_id:
            type: integer
            format: int32
          bus_license_number:
            type: string
          trip_date:
            type: string
            format: date
          first_name:
            type: string
          last_name:
            type: string
          user_email:
            type: string
            format: email

  TripUpdateAPIResponse:
    type: object
    properties:
      status:
        type: string
      error:
        type: string
securityDefinitions:
  #  UserPermission:
  #    type: basic
  #    description: |
  #      Roles required: Admin, Specialist
  Bearer:
    description: |
      For accessing the API a valid JWT token must be passed in all the queries in
      the 'Authorization' header.


      A valid JWT token is generated by the API and returned as answer of a call
      to the route /auth/login giving a valid user & password.


      The following syntax must be used in the 'Authorization' header :

          Bearer: xxxxxx.yyyyyyy.zzzzzz

    name: Authorization
    in: header
    required: true
    type: apiKey

paths:
  #######################################################################
  #  /examples:
  #    get:
  #      tags:
  #        - Examples
  #      description: Fetch all examples
  #      responses:
  #        '200':
  #          description: Returns all examples
  #    post:
  #      tags:
  #        - Examples
  #      description: Create a new example
  #      parameters:
  #        - name: example
  #          in: body
  #          description: an example
  #          required: true
  #          schema:
  #            $ref: #/definitions/ExampleBody
  #      responses:
  #        '200':
  #          description: Returns all examples
  #
  #    /examples/{id}:
  #      get:
  #        tags:
  #          - Examples
  #        parameters:
  #          - name: id
  #            in: path
  #            required: true
  #            description: The id of the example to retrieve
  #            type: integer
  #        responses:
  #          '200':
  #            description: Return the example with the specified id
  #          '404':
  #            description: Example not found
  #
  #    /spec:
  #      get:
  #        tags:
  #          - Specification
  #        responses:
  #          '200':
  #            description: Return the API specification
  #######################################################################
  /auth/signup:
    post:
      tags:
        - Authentication
      description: Register a new user
      summary: User can sign up
      operationId: signupUser
      produces:
        - application/json
      parameters:
        - name: register_body
          in: body
          description: User create object
          required: true
          schema:
            $ref: "#/definitions/UserRegisterBody"
      responses:
        '201':
          description: User got created
          schema:
            $ref: "#/definitions/AuthAPIResponse"
        '500':
          description: An internal server error occurs
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '409':
          description: Indentical user exists
          schema:
            $ref: "#/definitions/ErrorAPIResponse"

  /auth/signin:
    post:
      tags:
        - Authentication
      description: Log a user in
      summary: User can sign in
      operationId: signinUser
      produces:
        - application/json
      parameters:
        - name: login_body
          in: body
          description: The email/password
          required: true
          schema:
            $ref: "#/definitions/UserLoginBody"
      responses:
        '302':
          description: User got created
          schema:
            $ref: "#/definitions/AuthAPIResponse"
        '401':
          description: User is not authorized the access
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '404':
          description: No user with the prided credentials exists
          schema:
            $ref: "#/definitions/ErrorAPIResponse"

  /trips:
    get:
      tags:
        - Trips
      security:
        - Bearer: []
      description: Fetch all trips
      summary: Both Admin and Users can see all trips
      operationId: showTrips
      produces:
        - application/json
      responses:
        '200':
          description: Returns all trips
          schema:
            type: array
            items:
              $ref: "#/definitions/TripAPIResponse"
        '404':
          description: Fails to find any trip
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '403':
          description: Doesn't have enought permission to resources
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
    post:
      tags:
        - Trips
      security:
        #        - UserPermission:
        #            - admin
        #            - user
        - Bearer: []
      description: Creates a new trip
      summary: Admin can create a trip
      operationId: createTip
      produces:
        - application/json
      parameters:
        - name: tripCreateObject
          in: body
          description: Trip create object
          required: true
          schema:
            $ref: "#/definitions/TripBody"
      responses:
        '201':
          description: Trip got created
          schema:
            $ref: "#/definitions/TripAPIResponse"
        '500':
          description: An internal server error occurs
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '409':
          description: Indentical trip exists
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '403':
          description: Doesn't have enought permission to resources
          schema:
            $ref: "#/definitions/ErrorAPIResponse"

  /trips/{tripId}/cancel:
    patch:
      tags:
        - Trips
      security:
        - Bearer: []
      description: Cancel a specied trip
      summary: Admin can cancel a trip
      operationId: cancelTrip
      produces:
        - application/json
      parameters:
        - name: tripId
          in: path
          required: true
          description: The id of the trip to be cancelled
          schema:
            type: integer
          format: int32
          example: 1
          minimum: 1
      responses:
        '204':
          description: Trip with specific id got cancelled
          schema:
            $ref: "#/definitions/TripUpdateAPIResponse"
        '500':
          description: An internal server error occurs
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '404':
          description: No trip with the specified id exists
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '403':
          description: Doesn't have enought permission to resources
          schema:
            $ref: "#/definitions/ErrorAPIResponse"

  /trips/{tripId}:
    get:
      tags:
        - Trips
      security:
        - Bearer: []
      description: Retrieves a trips
      summary: Both Admin and Users can see a specific trip
      operationId: showTrip
      produces:
        - application/json
      parameters:
        - name: tripId
          in: path
          required: true
          description: The id of the trip to be retrived
          schema:
            type: integer
          format: int32
          example: 1
          minimum: 1
      responses:
        '200':
          description: Returns a specified trip
          schema:
            $ref: "#/definitions/TripAPIResponse"
        '404':
          description: Fails to find a trip with the specified id
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '403':
          description: Doesn't have enought permission to resources
          schema:
            $ref: "#/definitions/ErrorAPIResponse"

  /bookings:
    get:
      tags:
        - Bookings
      security:
        - Bearer: []
      description: Fetch all bookings (for Admin) or owned bookings (for User)
      summary: View all bookings. Admin can see all bookings while User can see all of his/her bookings
      operationId: showBookings
      produces:
        - application/json
      responses:
        '200':
          description: Returns all (for Admin) or owned by current user (for User) trip bookings
          schema:
            type: array
            items:
              $ref: "#/definitions/BookingAPIResponse"
        '404':
          description: No trip booking exists
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '403':
          description: Doesn't have enought permission to resources
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
    post:
      tags:
        - Bookings
      security:
        - Bearer: []
      description: Books a seat on trip
      summary: Users can book a seat on a trip
      operationId: bookTrip
      produces:
        - application/json
      parameters:
        - name: bookingCreateObject
          in: body
          description: Booking create object
          required: true
          schema:
            $ref: "#/definitions/BookingBody"
      responses:
        '201':
          description: Trip booking was created
          schema:
            $ref: "#/definitions/BookingAPIResponse"
        '500':
          description: An internal server error occurs
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '409':
          description: Indentical trip bookiing exists
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '403':
          description: Doesn't have enought permission to resources
          schema:
            $ref: "#/definitions/ErrorAPIResponse"

  /bookings/{bookingId}:
    delete:
      tags:
        - Bookings
      security:
        - Bearer: []
      description: Removes a seat booking from user trip booking list
      summary: Users can delete their bookings
      operationId: deleteBooking
      produces:
        - application/json
      parameters:
        - name: bookingId
          in: path
          required: true
          description: The id of the trip booking to be deleted
          schema:
            type: integer
          format: int32
          example: 1
          minimum: 1
      responses:
        '204':
          description: Trip booking with specific id got deleted
          schema:
            $ref: "#/definitions/BookingDeleteAPIResponse"
        '500':
          description: An internal server error occurs
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '404':
          description: No trip booking with the specified id exists
          schema:
            $ref: "#/definitions/ErrorAPIResponse"
        '403':
          description: Doesn't have enought permission to resources
          schema:
            $ref: "#/definitions/ErrorAPIResponse"